[I2C] Harden error management
Example for I2C scanner fewer errors are reported.
Sometime there is a HAL_I2C_ERROR_TIMEOUT was raised but
not reported as I2C_TIMEOUT.
In a general way, mask have to be used to properly handle error cases.

+++ libraries/Wire/src/Wire.cpp
@@ -247,7 +247,7 @@ uint8_t TwoWire::endTransmission(uint8_t sendStop)
 //
 uint8_t TwoWire::endTransmission(void)
 {
-  return endTransmission(true);
+  return endTransmission((uint8_t)true);
 }
 
 // must be called in:
